requirement.txt
google-generativeai
streamlit
python-dotenv
langchain
PyPDF2
chromadb